requirements.txt
aioboto3==12.2.0
aiobotocore==2.9.0
aiohttp==3.9.3
aioitertools==0.11.0
aiosignal==1.3.1
albumentations==1.3.1
altair==5.2.0
annotated-types==0.6.0
asttokens==2.4.1
attrs==23.2.0
blinker==1.7.0
boto3==1.33.13
botocore==1.33.13
cachetools==5.3.2
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.10.0
deeplake==3.8.17
defusedxml==0.7.1
dill==0.3.8
executing==2.0.1
Flask==3.0.1
fonttools==4.47.2
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.41
gunicorn==21.2.0
humbug==0.3.2
icecream==2.1.3
idna==2.10
imageio==2.33.1
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.3
libdeeplake==0.0.98
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib==3.8.2
mdurl==0.1.2
multidict==6.0.4
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.3
opencv-python-headless==4.8.0.74
packaging==23.2
pandas==2.2.0
pathos==0.3.2
pillow==10.2.0
pox==0.3.4
ppft==1.7.6.8
protobuf==4.25.2
pyarrow==15.0.0
pydantic==2.6.0
pydantic_core==2.16.1
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2023.3.post1
PyYAML==6.0.1
qudida==0.0.4
referencing==0.32.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.0
roboflow==1.1.19
rpds-py==0.17.1
s3transfer==0.8.2
scikit-image==0.22.0
scikit-learn==1.4.0
scipy==1.12.0
six==1.16.0
smmap==5.0.1
streamlit==1.31.0
supervision==0.18.0
tenacity==8.2.3
threadpoolctl==3.2.0
tifffile==2024.1.30
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.0.7
validators==0.22.0
waitress==3.0.0
watchdog==3.0.0
Werkzeug==3.0.1
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
